[Feat] 충돌 및 에러 해소
