+++ src/main/java/verbly/spring/domain/review/controller/ReviewController.java
@@ -1,7 +1,7 @@
 package verbly.spring.domain.review.controller;
 
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Null;
+import java.lang.Void;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -29,7 +29,7 @@ public class ReviewController {
 //            description = "JWT 인증된 유저가 nativeLang과 learningLang을 등록하는 API입니다.",
 //            security = @SecurityRequirement(name = "JWT TOKEN")
 //    )
-    public ResponseEntity<ApiResponse<Null>> createReview(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long revieweeId, @RequestBody @Valid ReviewRequestDTO reviewRequestDTO) {
+    public ResponseEntity<ApiResponse<Void>> createReview(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long revieweeId, @RequestBody @Valid ReviewRequestDTO reviewRequestDTO) {
 
         Long reviewerId = customUserDetails.getUser().getId();
         reviewService.createReview(reviewerId, revieweeId, reviewRequestDTO);
